New optimized ustring implementation of make_unique
This new implementation uses a custom data structure specific to this use case. It avoids re-hashing the string multiple times redundantly and uses memory pools to reduce the cost of allocations. It maintains the current strategy of using a binned hash table for concurrency, but uses rw spin locks to improve scalability. In my benchmarks, this new implementation is 2x to 3x faster.