oiiotool: bug fix - input file data format didn't always end up in output.
A semi-random assortment of mostly recently-added operations were calling
imagerec->read() instead of ot.read(imagerec). The confusing thing was that
the ot wrapper is where it was recording the tiling and pixel data format
of the first image read, for later use as the default for any output files
(in the absence of explicit -tile or -d directives). This would result in
output files with the "wrong" data formats.