IBA::type_merge utility
This encapsulates the heuristic about what pixel data type should be
used to combine two other data types, without loss of precision or
range. Roughly speaking, this means picking which of the two types (as
arguments) is more precise. But for some combinations there is no single
best answer, so we just punt and say a float is necessary.