DISPATCH_TYPES utility macro overhaul.
Background info:

The DISPATCH_TYPES macros are used internally to ImageBufAlgo to sort
out the specializations for the 9 different potential pixel data types
(uint8, int8, uint16, int16, uint32, int32, half, float, double).  Some
IBA functions really need to handle every combination without
compromise, but for many others it's too much code bloat to handle the
full cross-product of types (9x9x9 specializations for a function with
two input images and a result). For those cases, we have
DISPATCH_COMMON_TYPES_* that specialize for only the four common cases
(uint8, uint16, half, float) and *previously* issued an error if it was
one of the rare remaining types (e.g. it's exceedingly rare to come
across a signed 8 bit integer image in the wild).

OK, so this patch overhauls the DISPATCH macros, and in particular makes
the DISPATCH_COMMON ones fully robust to all data types without any
addition bloat. It dispatches the common cases as usual, but for the
uncommon ones, instead of aborting with an error, it just makes a
temporary FLOAT image copy of the image of oddball type, and falls back
to the FLOAT case.  Thus, we now will guarantee that all IBA functions
are expected to work for all pixel data types (though at the cost of an
extra image copy, for those rarely used types).

This simplifies the comments and documentation, which no longer have to
confusingly point out which functions work with all data types and which
work with only a subset.

Additionally, we add a new IBAPrep flag: IBAprep_CLAMP_MUTUAL_NCHANNELS.
If set, automatically clamps roi.chend to the maximum of that of the
inputs.  This improves safety and reduces separate checking logic for
the many IBA functions whose results would be meaningless or worse for
images of differing numbers of channels in the inputs. Once again, we
can simplify a lot of the documentation comments which previously had to
point out which functions were illegal to call with image arguments that
did not match in their number of channels. Now, it's almost always
legal, though for many cases if the channels don't match, the operation
will only be performed on those channels that they have in common.