Add destroy() methods to ImageInput and ImageOutput.
This is helpful on Windows if OIIO is used within plugins (DLLs) that may
be loaded and unloaded. It can be problematic to have memory allocated (as
is done by ImageOutput::create()) on one side of a DLL boundary, and then
freed on the other side of the DLL boundary (as would happen with an ordinary
'delete'). Instead, prefer to get rid of ImageInput*'s and ImageOutput*'s by
calling destroy().