IBA::colorconvert(): make it properly multithreaded.
We skipped this infrequently-used (at that time) functionality when we
were figuring out how to multithread all the IBA functions. Well, nowadays
it's used all the time, so a little code rearrangement allows it to use
all the threads. This makes color conversion via oiiotool go about 8x
faster on my workstation.