UTF-8 support for render_text algo
UTF8 decoder is from [here][1], it is a simple and fast decode, and will
take some invalid utf-8 input. It should not be a problem, since
FreeType will most likely not find such glyph. And it does not affect
program stability.

It is still up to user to supply font which supports all the needed
codepoints.

[1](http://bjoern.hoehrmann.de/utf-8/decoder/dfa/)