filesystem_test: make it pass for Boost 1.57
There is a spot where we check that Filesystem::scan_for_matching_filenames
doesn't crash with a wildly malformed regular pattern. We should merely
have called and made sure it returned. But we were checking that it
returned false. Problem is, it used to return false, but in Boost 1.57,
the malformed regex merely didn't match, rather than throwing an exception,
and that caused our scan_for_matching_filenames to return true rather
than false.