Moved utf8 code to strutil.h/cpp
Since strutil.h is public API, only utf8_to_unicode function is exposed.
It takes std::vector<int> as argument instead of returning it  so that
user can reserve space as needed.

This could in theory return u32string, but support for it is not yet
stabilized in c++11 compilers. And detecting such features is too much
work for simple utility function used only once (so far) in oiio itself.