Make TextureSystem::create(shared=true) truly create a shared TS.
It had been creating a separate TS for every create call, but with a
shared ImageCache (if requested).  That's not good enough. If two spots
in an app each request a shared TS, They probably want it to actually be
the same one, including any attributes already set in it.