oiiotool: give a proper error message when deep inputs are not supported.
Currently, most oiiotool operations are simply not supported for deep
inputs. But they aren't even smart enough to know that.

This patch requires a flag to be passed to IBAprep when deep inputs are
allowed; when not provided but images are deep, give a sensible error
message.  This is much better than crashing!