oiiotool -crop : fix bug in the logic
Nobody reported this as a problem, but in looking at action_crop, I
think the logic was all wrong. If the crop region was the same size as
the original pixel data window, it would merely change the origin. That
can't be right (and, in any case, is accomplished with the separate
--origin command).