Include some Droid fonts (from the Android SDK, Apache-licensed)
This ensures that we have one sans, one serif, and one mono font that will
always be available and always identical on all platforms. At the very least,
this makes unit tests a lot easier, by decoupling them from the whims of whatever
random set of system fonts happen to be installed.

Use DroidSans as the default font when none is specified.
Touch up the font searching logic to increase the chances of it getting found
wherever the OIIO installation winds up.