oiiotool -n and --debug
-n : "no saved output" performs all the calculations (including timing
and stats) but does not write any output files to disk. This is helpful
for debugging, for testing an oiiotool command line syntax or behavior
without wanting to accidentally write files with a malformed command,
and for performance testing of computation where you don't really care
about timing of I/O to write the tiles.

--debug : Previously the "-v" (verbose) performed two somewhat unrelated
tasks -- it increased the verbosity of essential output, but it alto
triggered certain debugging status messages. Now we have separate
controls for these two options.