psdinput: better error handling for files lacking additional layer info
Some PSDs don't have a valid "Global layer mask info" or "Addtional layer info"
subsection in the "Layer and Mask Infomation" section of the file format causing
parse errors. Detecting this situation and seeking to the beginning of the next
section allows OIIO to support a greater range of PSD files.

Signed-off-by: Matthew E. Levine <melevine@gmail.com>