ustring fixes for new gcc (5.1+) std::string ABI
This makes it work (and adds a test).
We still need to return and do something more clever for that case.
I'll need to install gcc 5.1 first to see how the internals work.
But in the mean time, at least this will unbreak things for people
already on gcc 5.1.

Fixes #1174