Change Python ImageOutput bindings to simplify the write_* methods.
They, oddly, took both an array and a TypeDesc describing the array.
Nix the TypeDesc and just figure out the array type. If they didn't
match, it would only have been trouble. Also, get rid of the stride
parameters, they make sense in a C++ context, but are just confusing to
Python.