Fix bugs with rangecompress and rangeexpand when using luma.
This fixes a bug where the (potentially uninitialized) destination buffer was used as the source for computing luma, when the destination and source buffers are different. Also adds a check against invalid scaling due to zero or negative luma (which might appear as residual error from a previous computation), and additional cases for the test suite.