Have TIFF output recognize special "tiff:write_exif" metadata.
If present and set to 0, will not write an Exif tag directory to the
file, even if Exif metadata is otherwise present. This can be useful
when writing a TIFF file intended to be read by an application that
links to a version of libtiff that is too old to properly read the Exif
data and don't want it confused.

Note in particular that PRMan falls into that category. Therefore, this
patch also modifies 'maketx' so that when you use the --prman flag, it
additioally does the equivalent of -attrib "tiff:write_exif" 0