PTex build changes -- use external PTex if available.
Using FindPTex.cmake from the PTex project itself, search for an external
PTex library. Use it if found.

If external PTex is not found, and as long as USE_PTEX is not disabled,
fall back on the existing "embedded source" PTex implementation.

I will eventually remove the embedded one entirely, once I get the
all-clear from everybody (including my facility, which needs it for a
little while longer).