Make TIFF read robust to subimge-to-subimage changes in some header fields.
Rearrange the order in which we examine TIFF header order. When reading
MIPmap levels (beyond the first one), we assume that all the ordinary
metadata should be the same for all levels, so we don't laboriously read
all that stuff. But in our zeal to skip this needless work, we took the
early out before examining some essential fields such as planarconfig,
photometric interpretation, palette processing, and extrasamples.  If a
file should change this from one MIP level to the next -- which no sane
file should do, but hey, it did happen in the wild -- that would bring
chaos. Now we are more robust to this, while still avoiding all the
detailed metadata reading for non-head MIP levels.

Fixes #1220