Allow ImageCache/TextureSystem to store partial channel sets in cache.
ImageCache's original design would read all channels into a cached tile,
because (1) most files have only a few channels, (2) most uses (for
example, texture) tend to use most or all of the channels in the file,
in close temporal proximity if not in the same lookup, (3) most file
formats don't have a way to read a channel subset that's any more
efficient than reading all of them.

However, for the rare case that there are MANY channels, only a few of
which are needed (for example, looking up 3 channels of textures from a
40 channel file), this is very wasteful of texture cache space, and MAY
(depending on the format) also be wasteful of I/O.

So this patch:

* Modifies ImageCache so that a cached tile can store any contiguous
  subset of channels. This means the TileID stores a chbegin/chend
  range, the hashes consider this, and certain ImageCache public API
  calls now have varieties that take chbegin/chend parameters.

* ImageCache will create cached tiles that store exactly the channel
  range requested. The default is all channels in the file. Use subsets
  wisely -- if you have a 4-channel file and you separately look up
  channel ranges [0,3) and [0,4), it will redundantly read and store
  the overlapping channels!

* TextureSystem is modified to ask for all channels when there are only
  a few channels in the file, but when there are more than
  max_tile_channels (a new settable attribute of the TS) channels in
  the file, it will ask the IC for only the channels it needs for the
  texture lookup. The max_tile_channels is set by default to 5, which
  means for the vast majority of files (R, RGB, RGBA, RGBAZ) it will
  read all channels and behave as before. But if you have a mega-channel
  file, it will try to read and cache only the subset of channels you
  requested with the texture call. If you don't want the new behavior,
  just set max_tile_channels to something bigger.

The bottom line is that if your app uses IC with the old API calls, or
if you use TS with texture files having 5 or fewer channels (which is
almost always the case), you should see no change in behavior or
performance. But for those of you who for some reason deal with files of
many many channels, accessing a small subset of them via texture or IC
should result much more efficient cache behavior because the IC won't be
caching channels you aren't accessing.