Separate global "threads" and "exr_threads" attributes.
The global OIIO "threads" attribute controls the default "fan out" of
spawned threads when an OIIO call has work that can be parallelized
(like for ImageBufAlgo functions). Apps with essentially single logic
thread leave this at 0, which means to spawn #cores threads, whereas
apps with lots of app-side threads set it to 1 to prevent OIIO from
internally spawning many threads for each app calling thread.

Unfortunately, we were using the same attribute value to set the size of
OpenEXR's thread pool size. In an app where we set "threads" to 1 to
prevent OIIO fan-out, this was unfortunately also limiting OpenEXR's
worker thread pool to 1, which means the various threads would block
against each other.

It seems like the best solution is to break this apart into separate
attributes: "threads" to control the OIIO fan-out, and "exr_threads"
to control the size of OpenEXR's thread pool. Most apps need not mess
with either of these.