Fix recent exr metadata break
A few checkins ago, we refactored some EXR metadata output, making a
table.  Well, turns out that this can be subject to the "initialization
order fiasco", because the static table in exroutput.cpp referenced the
static global variables TypeDesc::TypeMatrix (and several others) that
are defined in typedesc.cpp, and thus may not yet have been initialized.

Why, exactly, the fairly trivial constructors don't automatically bake
down into POD that can be statically initialized in the DSO, I don't
understand.

But anyway, slight modification of the table to not depend on any prior
initializations fixes it.

Also put in automatic float[16]-to-matix44 translation, though it turns
out that wasn't the problem after all, and beefed up a debug message.