ImageInput: Add nthreads parameter to read_image, read_scanlines, read_tiles.
This allows control over thread spawning, similar to the parameter on
the various ImageBufAlgo functions. A value of 0 means to use the global
OIIO::attribute("nthreads") value (which in turn defaults to have a
thread fan-out equal to the hardware concurrency). For a sole thread of
execution, you want a big fan-out for tasks that can be parallelized,
whereas when the calling function is one of many concurrent threads
already underway, you just want the work done directly by the calling
thread, without spawning additional threads that will oversubscribe the
hardware.

There are currently only a couple spots where ImageInput might spawn
threads. So this is primarily for future expansion, so that we might add
more multithreading and give appropriate control to the calling
application.

The reason that we want per-call overrides rather than relying solely on
the global "nthreads" attribute is that OIIO may be used by multiple
parts of an application (or multiple dependent libraries or plugins,
each unaware of the others), and so it is not particularly safe under
those circumstances to set the global value and assume that it will stay
that way.

This is a preliminary effort for design review. Obvious extensions, if
people like the way this is going, would be to add the nthreads
parameter to the "native" versions (read_native_tiles,
read_native_scanlines), the "deep" versions, and and corresponding
routines in ImageOutput. Also we should add a method to ImageBuf that
can set an "nthreads" paremeter for any underlying ImageInput or other
operations done internal to the IB.

It is worth discussing alternatives. I admit that the idea of scattering
"nthreads" parameters widely through the APIs is one that does give me
some pause. But it has worked well for IBA, and at the moment I don't
have a better idea for ImageInput/ImageOutput. Feedback appreciated.