API changes to allow thread policy per ImageInput, ImageOutput, and ImageBuf.
This allows control over thread spawning for any ImageInput,
ImageOutput, and ImageBuf internals that may have the opportunity to
parallelize work, similar to the parameter on the various ImageBufAlgo
functions. A value of 0 means to use the global "threads" attribute
value (which in turn defaults to have a thread fan-out maximum equal to
the hardware concurrency). For a sole thread of execution, you want a
big fan-out for tasks that can be parallelized, whereas when the calling
function is one of many concurrent threads already underway, you just
want the work done directly by the calling thread, without spawning
additional threads that will oversubscribe the hardware (i.e., setting
this value to 1).

The reason that we want per-ImageInput/Output/Buf overrides rather
than relying solely on the global "threads" attribute is that OIIO may
be used by multiple parts of an application (or multiple dependent
libraries or plugins, each unaware of the others), and so it is not
particularly safe under those circumstances to set the global value and
assume that it will stay that way. It depends a lot on the app, and this
patches gives apps more flexibiity in determining threading policy.