Overhaul of DeepData internals and minor change to API
Internally, change the data storage order from [pix][chan][samp] to
[pix][samp][chan], so that now all the channels of one sample are
contiguous (as we store them internally for flat images).

We should have done it this way all along. I thought the other order was
required for OpeneEXR (based on a misreading of the example code), but
it turns out that proper use of strides lets us store the channels
interleaved in memory if we want.

No longer store pointers internally at all, just need to keep the prefix
sum of the number of samples for each pixel, in order to quickly compute
the data offset of a particular sample or channel.  When we need
pointers "externally" (currently, only for read/write of OpenEXR), use a
new get_pointers() method that calculates them on the fly.

This lets us add insert_samples() and erase_samples(), which let us
(simply, if not cheaply) alter the samples in an existing DeepData, which
will be the basis of much more extensive support for deep files in
ImageBufAlgo and oiiotool (which support deep files now in only an
extremely minimal set of operations).

Get rid of DD::channel_ptr(), which no longer makes a lot of sense, and
instead have data_ptr(). Use with caution, because the pointers will no
longer be valid if anything causes the internal data buffer to
reallocate.

Get rid of the DD::alloc() method. We no longer require client code to
trigger the allocation of the data array. Instead, it will automatically
happen the first time you read or write a pixel value if it's not
already allocated.