Fix for compilation with Boost 1.60 -- major Python ImageCache overhaul
I don't know why it worked before or why it's broken now, but the
just-release Boost 1.60 refused to build OIIO, and all the problems
were in the ImageCache bindings. And they were weird, and in retrospect
some of the calls would probably not ever have worked. So after a big
refactor, we are building against Boost 1.60 and passing all tests.