Testsuite organization overhaul -- no more src or ref in build
The old way was that when CMake runs for the first time (when it builds
your make or ninja files), it copies everything from testsuite/ into
build/PLATFORM/testsuite/, and then when you run tests, it uses all the
info (source images, scripts, and reference output) in the build area.

That was fine usually, but if you were developing or modifying a test
itself, it was a big PITA, because changing something in
testsuite/mytest would not be reflected in the
build/PLATFORM/testsuite/mytest that had already been copied. It was the
source of many frustrations and errors. And beware modifying something
in the build area and then blowing it away, not remembering to copy it
to the source testsuite/ area.

This overhaul NO LONGER COPIES source data or reference output to build at
all. Instead, a combination of links and references to the canonical files
are used to make the build area only the temporary location of test output
itself. This sure does make everything easier.