Double speed of ImageBuf read by bypassing the IC for "forced" reads.
IB was usually falling back on IC::get_pixels to go through the cache,
even when asked to "force" a read. Making it accept a wider variety of
circumstances in which it just does a direct read via
ImageInput::read_image (and not making an extra copy out of the IC)
roughly doubles read performance for typical cases.

I suppose that this could bite you if you ended up with a bunch of IB's
in the app all wanting the same image, in which case going through the
cache would have been a win. And this was obviously the intent of the
original design. But I think we have enough experience now to realize
that this is an unusual case -- in most apps (aside from "texture") a
particular image is not likely to be read more than once in a short
period of time, and so does not benefit from a copy being kept in an
ImageCache. So it's probably worth the tradeoff to get higher read
performance everywhere else that IB's are used.

If an app thinks it will read the same images over and over, perhaps it
should not pass force=true for IB::read() calls, and instead more
intentionally rely on the IC.