Change default to USE_CPP11=1     #1344
This does NOT make the codebase C++11-only. Although that will happen
sooner or later, almost certainly during the course of 1.7 development.
But for now, this just means that it defaults to C++11, and you need to
proactively use 'make USE_CPP11=0' (or 'cmake -DOIIO_BUILD_CPP11=0') to
stay on C++03 compatibility. This is to ensure that nobody is surprised
when the C++11 switch eventually comes, by requiring them to have
continued awareness that they are living on borrowed time.