oiiotool --attrib:type=... name value
Optional type=typename modifier of the --attrib lets you be explicit
about the type. This is helpful in two situations: (1) avoid ambiguity
of the usual rules that deduce the type from the value, for example if
you want to set a string to "0" without being misinterpreted as an
integer. (2) use an unusual or aggregate type, such as setting a
camera matrix with

    oiiotool in.exr --attrib:type=matrix worldtocam 1,0,0,0,.... -o out.exr