oiiotool -otex, -oenv writes out texture or env maps directly from oiiotool No need to run maketx separately on oiiotool output.
There's no particular need to replace your current use of maketx with
oiiotool, it's not planned to go away. But there are times when you use
oiiotool to do a variety of operations to a texture (compositing, color
adjustment, shuffling channels, etc.) and now you can write those
adjusted images out directly as textures.

Thus, generally speaking,

    maketx in.tif -o out.tx

should be equivalent to

    oiiotool in.tif -otex out.tx

But how to handle all of the possible command-line arguments of maketx?

Some are identical for oiiotool as they were for maketx:
    -v  --threads --runstats  --chnames  -d
    --tile --compression --fixnan  --fullpixels
    --attrib --sattrib  --sansattrib

Some have equivalencies in oiiotool, but are slightly altered due to
existing command line syntax differences between the two programs:

    maketx               oiiotool
    --nchannels N    ==  --ch 0,1,...N-1
    --separate       ==  --planarconfig separate
    --fovcot X       ==  --attrib fovcot X
    --Mcamera        ==  --attrib:type=matrix worldtocamera ...
    --Mscreen        ==  --attrib:type=matrix worldtoscreen ...

But a whole lot can be set as options to the -otex command, for
example:

    maketx in.tif -wrap periodic -opaque-detect -o out.tx

is the same as

    oiiotool in.tif -otex:wrap=periodic:opaque_detect=1 out.tx