First stab at code coverage
make CODECOV=1
make CODECOV=1 test

on a gcc/clang system will run the testsuite and generate a code coverage
report in build/$PLATFORM/cov/index.html

I'm not yet sure how we should use this. It's just for information and
to help spot places in the code that aren't covered by the
testsuite. But I do not think we should aim to get 100% coverage --
eliminating all the stragglers for every possible error case and early
return would be so much labor that we'd never get any real work done. So
take with a grain of salt and use wisely.