simd exp, log (exact and approximate/fast versions)
fast_pow_pow is a simd approximate pow that is only designed to handle
positive values. But for that case, is even faster for not needing to
consider other cases. Very handy for pow in cases where you know that
the range of values would be safe, such as sRGB->linear conversion.

Also add code to simd_test to not only verify that they work, but also
benchmark them.