oiiotool -native behavior adjustment
This actually brings --native more in line with what the documentation
has said it does.

Old behavior: By default, images read from disk are backed by ImageCache
(with "forcefloat" on) unless they are modestly sized (in which case
they bypass the IC for speed, but are still converted to float). The
--native option caused image reads to totally bypasses the ImageCache
(direct into an ImageBuf, even for small images) and also stay as close
as possible to their native format.

New behavior: Same default, but now --native option turns off
"forcefloat" for the ImageCache, keeps the data as close as possible to
its native format. It will still use the cache as backing for big images
if it's one of the common types directly supported by the IC (float,
half, uint16, uint8), but otherwise will read directly into an ImageBuf.

If you're doing any image math, it's best (faster and more accurate) to
use the default behavior of conversion to float upon input. But if
you're basically just using oiiotool for file format or data format
conversion, you can save time and memory by sticking to the native
format. There are cases where this is handy, but only if you know what
you are doing.