oiiotool: add "-i" command for input. (1389)
It has always been that simply naming a file causes it to be an input.
In its basic form, -i is simply a synonym. But it can also be used as
an anchor for per-file options. Start with autocc=, now=, info=.

The "now" option is particularly interesting (and the reason why I added
this now) because if you are benchmarking and looking at results with
--runstats, ordinarily the reads are done lazily and so time spent
actually reading the file is unclear and spread into other functions
that need the data. By using -i:now=1 you can force the input read time
to be accounted separately.