Use libjpeg-turbo if found, roughly doubles JPEG I/O perf. (1390)
This is a drop-in replacement for libjpeg. In my tests, I'm finding it
around 1.7x faster for reading JPEG files, and 2.2x faster for writing
JPEG files. This is on my laptop with SSD, so it may be less pronounced
with a slow disk or network where the actual transfer of the bytes
is more significantly costly than the encode/decode.

It's on by default (IF it finds it on your system, otherwise it falls back
to the usual libjpeg). But you can turn it off with USE_JPEGTURBO=0.

It can exhibit some minor LSB-type differences compared to libjpeg. I
have not yet seen any visible artifacts from this. The differences are
minor enough that it seems well worth a 2x performance improvement for
JPEG decoding.