WIP Support random access writes to subimages/MIPmaps.
This adds to ImageOutput a new version of write_scanlines() and
write_tiles() that takes subimage and miplevel designations for formats
that support true random access to subimages/miplevels without the need
to write a subimage or miplevel to completion before moving on to another.

We use supports("random_subimage_access") to designate ImageOutput
plugins that support this.

The presumed use case is for a renderer that is generating many AOVs
simultaneously, each one to a separate subimage in the same file (in
OpenEXR parlance, to different "parts") and as it renders scanlines or
buckets (tiles), wants to write all the AOVs for the current bucket to
their respective subimage, without needing to buffer all the AOVs for
the entire image.

Most ImageOutput implementations are not expected to support this
feature.  In fact, the only format I'm aware of that is capable of
supporting it is OpenEXR.

As of this checkin, the OpenEXR support as not yet been added. This is
just a conceptual prototype for review of the API.