Move the few stray find_package calls to externalpackages.cmake.
The idea is to make sure that after including externalpackages.cmake,
we are 100% confident that all the FOO_FOUND symbols are set, or by not
being set correctly reflect which dependencies were not found.

The end goal (of which this is just an organizational prerequisite) is
to some day augment the build system with an option to download and
build all dependencies that are not found on the system. A first step is
to have a specific known point (after inclusion of
externalpackages.cmake) where you know for sure which packages are
present and which are not, rather than have that logic dribbled across a
bunch of CMakeLists files in various directories.